+++ pyimouapi/const.py
@@ -23,6 +23,10 @@ API_ENDPOINT_GET_PRODUCT_MODEL = "/openapi/getProductModel"
 API_ENDPOINT_GET_IOT_DEVICE_DETAIL_INFO = "/openapi/getIotDeviceDetailInfo"
 API_ENDPOINT_WAKE_UP_DEVICE = "/openapi/wakeUpDevice"
 API_ENDPOINT_SET_MESSAGE_CALLBACK = "/openapi/setMessageCallback"
+API_ENDPOINT_GET_COLLECTION = "/openapi/getCollection"
+API_ENDPOINT_TURN_COLLECTION = "/openapi/turnCollection"
+API_ENDPOINT_SIREN_START = "/openapi/sirenStart"
+API_ENDPOINT_SIREN_STOP = "/openapi/sirenStop"
 
 # error_codes
 ERROR_CODE_SUCCESS = "0"
@@ -109,6 +113,10 @@ PARAM_MOTION_DETECT = "motion_detect"
 PARAM_STORAGE_USED = "storage_used"
 PARAM_RESTART_DEVICE = "restart_device"
 PARAM_NIGHT_VISION_MODE = "night_vision_mode"
+PARAM_COLLECTION_POINT = "collection_point"
+PARAM_COLLECTION_POINT_PROMPT = "select_collection_point"
+PARAM_SIREN_START = "siren_start"
+PARAM_SIREN_STOP = "siren_stop"
 PARAM_PTZ = "ptz"
 PARAM_TEMPERATURE_CURRENT = "temperature_current"
 PARAM_HUMIDITY_CURRENT = "humidity_current"
@@ -128,6 +136,18 @@ PARAM_ABILITY_REFS = "abilityRefs"
 PARAM_REF_TYPE = "ref_type"
 PARAM_EXPRESSION = "expression"
 PARAM_OUTPUT_DATA = "outputData"
+PARAM_TURN_REF = "turn_ref"
+PARAM_TURN_INPUT_REF = "turn_input_ref"
+PARAM_INPUT_REF = "input_ref"
+PARAM_COLLECTION_NAME = "name"
+
+IOT_GET_COLLECTION_REF = "21500"
+IOT_TURN_COLLECTION_REF = "22000"
+IOT_TURN_COLLECTION_NAME_REF = "22001"
+IOT_COLLECTION_NAME_REF = "21551"
+IOT_SIREN_START_REF = "25500"
+IOT_SIREN_START_INPUT_REF = "25501"
+IOT_SIREN_STOP_REF = "22200"
 PARAM_VALUE_TYPE = "value_type"
 PARAM_ACCESS_TYPE = "accessType"
 PARAM_ABILITY = "ability"
@@ -294,6 +314,8 @@ SWITCH_TYPE_REF = {
 #  Required capacity for various button types
 BUTTON_TYPE_ABILITY = {
     "restart_device": ["Reboot"],
+    "siren_start": ["Siren"],
+    "siren_stop": ["Siren"],
     "ptz_up": ["PT", "PTZ"],
     "ptz_down": ["PT", "PTZ"],
     "ptz_left": ["PT", "PTZ"],
@@ -330,6 +352,15 @@ BUTTON_TYPE_REF = {
                 "35gL0U5A",
             ],
         },
+    ],
+    "siren_start": [
+        {
+            "ref": IOT_SIREN_START_REF,
+            "input_ref": IOT_SIREN_START_INPUT_REF,
+        },
+    ],
+    "siren_stop": [
+        {"ref": IOT_SIREN_STOP_REF},
     ],
     "ptz_up": [
         {"ref": "22100"},
@@ -371,6 +402,7 @@ BUTTON_TYPE_REF = {
 #  Required capacity for various select types
 SELECT_TYPE_ABILITY = {
     "night_vision_mode": ["NVM"],
+    "collection_point": ["CollectionPoint"],
 }
 SELECT_TYPE_REF = {
     "night_vision_mode": [
@@ -402,6 +434,16 @@ SELECT_TYPE_REF = {
             "default": "0",
             "options": ["99", "0", "1", "2"],
             "value_type": "int",
+        }
+    ],
+    "collection_point": [
+        {
+            "ref": IOT_GET_COLLECTION_REF,
+            "turn_ref": IOT_TURN_COLLECTION_REF,
+            "turn_input_ref": IOT_TURN_COLLECTION_NAME_REF,
+            "ref_type": "services",
+            "default": PARAM_COLLECTION_POINT_PROMPT,
+            "options": [PARAM_COLLECTION_POINT_PROMPT],
         }
     ],
 }